ExampleProjects/DecaMoveUnityDemo/Python/requirements.txt
pynput==1.7.3
pyzmq==19.0.0
